requirements.txt
celery
requests
Django
pandas
pillow
redis
billiard